изменил вывод расчетов на float чтоб было адекватное количество нулей после запятой.

+++ src/com/academy/lesson02/HomeWorkLesson02/Task02.java
@@ -5,13 +5,13 @@ import java.util.Scanner;
 public class Task02 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        int weightUser;
+        double weightUser;
         System.out.println("Program that calculates weight on the moon.");
-//использовано входящее число типа INT т.к. double оставляет нули после запятой.
+//поигрался с типами данных, чтоб на вход принимать Double, а выводить адекватное колличество нулей после запятой.
         System.out.print("Entre your weight: ");
-        weightUser = scanner.nextInt();
-        System.out.format("Your weight on Earth = %s kg.", weightUser);
-        System.out.format("\nYour weight on Moon = %s kg." , (weightUser * 0.17));
+        weightUser = scanner.nextDouble();
+        System.out.format("Your weight on Earth = %s kg.", ((int)(weightUser)));
+        System.out.format("\nYour weight on Moon = %s kg." , ((float)(weightUser * 0.17)));
     }
